pyproject.toml
[tool.poetry]
name = "UdeCursos Bot"
version = "2.1"
description = ""
authors = ["CxrlosKenobi <kenobi@ieee.org>", "AlbFR <AlbFR@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^20.0a0"
beautifulsoup4 = "^4.11.1"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
